Rewrites court api availability check
