+++ libs/judicial-system/court-client/src/lib/courtClient.module.ts
@@ -1,9 +1,18 @@
 import { Module } from '@nestjs/common'
 
 import { CourtClientService } from './courtClient.service'
+import { CourtClientAvailableService } from './courtClientAvailable.service'
+import { CourtClientUnavailableService } from './courtClientUnavailable.service'
+import { courtClientModuleConfig } from './courtClient.config'
 
+const courtClientProvider = {
+  provide: CourtClientService,
+  useClass: courtClientModuleConfig().courtApiAvailable
+    ? CourtClientAvailableService
+    : CourtClientUnavailableService,
+}
 @Module({
-  providers: [CourtClientService],
-  exports: [CourtClientService],
+  providers: [courtClientProvider],
+  exports: [courtClientProvider],
 })
 export class CourtClientModule {}
